pyproject.toml
[project]
name = "medical-healthcare-rag"
version = "0.1.0"
description = "Multimodal medical document RAG system with mandatory citations"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Orchestration
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    # Vector DB
    "weaviate-client>=4.9.0",
    # Embeddings (Google AI Studio — gemini-embedding-exp-03-07)
    "google-genai>=1.0.0",
    # LLM (Fireworks AI — OpenAI-compatible)
    "openai>=1.50.0",
    # Document loading
    "pymupdf>=1.24.0",
    "boto3>=1.34.0",
    # Tokenisation
    "tiktoken>=0.8.0",
    # Validation / config
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # HTTP
    "httpx>=0.27.0",
    # Observability
    "structlog>=24.4.0",
    # Retry
    "tenacity>=9.0.0",
    # Reranking
    "cohere>=5.0.0",
    "boto3>=1.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# ---------------------------------------------------------------------------
# Ruff
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

# ---------------------------------------------------------------------------
# Mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
check_untyped_defs = true

# ---------------------------------------------------------------------------
# Pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: requires live services (Weaviate, Vertex AI, Fireworks)",
]
